refactor(scripts): split docs route inventory by pages, reference, assembly
Extract hand-authored page/endpoint tables and schema-derived geom/stat/
position detail routes so inventory assembly and validation stay thin.
Register the new modules on the docs CI lane and content-hash surface so
metadata-only edits cannot false-green. Pin detail-route heading matchers
for alias and related-examples gates.
