+++ scripts/docs-route-inventory.ts
@@ -6,11 +6,17 @@ import {
   isInteractionExposition,
 } from "../apps/docs/src/lib/catalog/interaction-exposition.ts";
 import { GUIDE_CATALOG, type GuideCatalogEntry } from "../apps/docs/src/lib/catalog/guide.ts";
-import type { DocsRouteMetadata, RouteHeading } from "../apps/docs/src/lib/route-types.ts";
-import { geomReferenceList } from "../packages/spec/src/geom-reference.ts";
-import { positionReferenceList } from "../packages/spec/src/position-reference.ts";
-import { statReferenceList } from "../packages/spec/src/stat-reference.ts";
-import { CLI_REFERENCE_OPTIONS } from "./cli-docs.ts";
+import type { DocsRouteMetadata } from "../apps/docs/src/lib/route-types.ts";
+import {
+  ENDPOINT_ROUTES,
+  PERFORMANCE_ROUTES,
+  TOP_LEVEL_ROUTES,
+} from "./docs-route-inventory-pages.ts";
+import {
+  geomDetailRoutes,
+  positionDetailRoutes,
+  statDetailRoutes,
+} from "./docs-route-inventory-reference.ts";
 
 /** Script-side name for the shared route metadata contract (`DocsRouteMetadata`). */
 export type DocsRouteRecord = DocsRouteMetadata;
@@ -20,327 +26,6 @@ export type {
   RouteHeading,
   RouteNavigation,
 } from "../apps/docs/src/lib/route-types.ts";
-
-const TOP_LEVEL_ROUTES: readonly DocsRouteRecord[] = [
-  {
-    path: "/",
-    title: "ggsvelte — layered grammar of graphics for Svelte",
-    description:
-      "Layered grammar of graphics for Svelte: ggplot2-style aes, geoms, stats, and themes, with PortableSpec JSON and hybrid SVG/canvas rendering.",
-    canonicalPath: "/",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "site",
-  },
-  {
-    path: "/docs",
-    title: "Documentation — ggsvelte",
-    description: "Install, compose the grammar, add interaction, ship, and read public contracts.",
-    canonicalPath: "/docs",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Start", label: "Overview", order: 0 },
-  },
-  {
-    path: "/examples",
-    title: "Gallery — ggsvelte",
-    description: "Runnable ggsvelte examples across marks, stats, scales, and interaction.",
-    canonicalPath: "/examples",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "site",
-  },
-  {
-    path: "/themes",
-    title: "Chart themes — ggsvelte",
-    description: "Built-in chart themes for paper, grids, axes, and type.",
-    canonicalPath: "/themes",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "site",
-  },
-  {
-    path: "/palettes",
-    title: "Color palettes — ggsvelte",
-    description: "Categorical palettes and sequential color scales for data encoding.",
-    canonicalPath: "/palettes",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "site",
-  },
-  {
-    path: "/interactions",
-    title: "Chart-local interaction — ggsvelte",
-    description:
-      "Inspect, select, zoom, and legend focus on a live chart. Semantic state is opt-in.",
-    canonicalPath: "/interactions",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "site",
-  },
-  {
-    path: "/reference",
-    title: "Reference — ggsvelte",
-    description: "Geom, interaction, CLI, diagnostic, lifecycle, and PortableSpec contracts.",
-    canonicalPath: "/reference",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Reference", label: "Reference overview", order: 50 },
-  },
-  {
-    path: "/reference/geoms",
-    title: "Geoms — ggsvelte",
-    description:
-      "Schema-derived API reference for every Geom* component: defaults, allowed stats and positions, and params.",
-    canonicalPath: "/reference/geoms",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Reference", label: "Geoms", order: 51 },
-    headings: [
-      { id: "all-geoms", title: "All geoms", level: 2 },
-      { id: "shared-layer-props", title: "Shared layer props", level: 2 },
-    ],
-  },
-  {
-    path: "/reference/stats",
-    title: "Stats — ggsvelte",
-    description:
-      "Schema-derived API reference for every statistical transform: after_stat columns and compatible geoms.",
-    canonicalPath: "/reference/stats",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Reference", label: "Stats", order: 52 },
-    headings: [
-      { id: "all-stats", title: "All stats", level: 2 },
-      { id: "how-to-set", title: "How to set a stat", level: 2 },
-    ],
-  },
-  {
-    path: "/reference/positions",
-    title: "Positions — ggsvelte",
-    description:
-      "Schema-derived API reference for every position adjustment: positionParams and compatible geoms.",
-    canonicalPath: "/reference/positions",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Reference", label: "Positions", order: 53 },
-    headings: [
-      { id: "all-positions", title: "All positions", level: 2 },
-      { id: "how-to-set", title: "How to set a position", level: 2 },
-    ],
-  },
-  {
-    path: "/reference/interactions",
-    title: "Search interactions — ggsvelte",
-    description:
-      "Filter ggsvelte interaction capabilities, events, diagnostics, and accessibility defaults.",
-    canonicalPath: "/reference/interactions",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Reference", label: "Interaction reference", order: 54 },
-  },
-  {
-    path: "/reference/cli",
-    title: "Command-line reference — ggsvelte",
-    description:
-      "Render PortableSpec JSON to SVG with implementation-derived flags, streams, diagnostics, and exit classes.",
-    canonicalPath: "/reference/cli",
-    kind: "page",
-    index: true,
-    sitemap: true,
-    shell: "docs",
-    navigation: { section: "Reference", label: "CLI reference", order: 55 },
-    headings: [
-      { id: "input-and-output", title: "Input and output", level: 2 },
-      { id: "options", title: "Options", level: 2 },
-      ...CLI_REFERENCE_OPTIONS.map((option) => ({
-        id: option.anchor,
-        title: option.flag,
-        level: 3,
-      })),
-      { id: "exit-codes", title: "Exit codes", level: 2 },
-      { id: "troubleshooting", title: "Troubleshooting", level: 2 },
-    ],
-  },
-];
-
-/** Same matching rule as apps/docs reference/geoms/[name] page load. */
-function geomHasRelatedExamples(geom: string): boolean {
-  const compact = geom.replaceAll("_", "");
-  const dashed = geom.replaceAll("_", "-");
-  return EXAMPLES.some(
-    (entry) =>
-      entry.category === geom ||
-      entry.category === compact ||
-      entry.tags.includes(geom) ||
-      entry.tags.includes(dashed),
-  );
-}
-
-/** One indexable page per KNOWN_GEOMS entry, derived from SpecDeclarations. */
-function geomDetailRoutes(): DocsRouteRecord[] {
-  return geomReferenceList().map((entry) => {
-    const headings: RouteHeading[] = [{ id: "defaults", title: "Defaults", level: 2 }];
-    // Order matches apps/docs/src/routes/reference/geoms/[name]/+page.svelte.
-    if (entry.aliasOf !== undefined) {
-      headings.push({ id: "alias", title: "Alias", level: 2 });
-    }
-    headings.push(
-      { id: "svelte", title: "Svelte component", level: 2 },
-      { id: "json", title: "JSON layer", level: 2 },
-      { id: "params", title: "Params", level: 2 },
-      { id: "allowed-stats", title: "Allowed stats", level: 2 },
-      { id: "allowed-positions", title: "Allowed positions", level: 2 },
-    );
-    if (geomHasRelatedExamples(entry.name)) {
-      headings.push({ id: "examples", title: "Examples", level: 2 });
-    }
-    return {
-      path: `/reference/geoms/${entry.slug}`,
-      title: `${entry.component} — ggsvelte`,
-      // Prefix with component so indexable descriptions stay unique when layer
-      // summaries share phrasing across aliases or related marks.
-      description: `${entry.component}: ${entry.summary}`,
-      canonicalPath: `/reference/geoms/${entry.slug}`,
-      kind: "page" as const,
-      index: true,
-      sitemap: true,
-      shell: "docs" as const,
-      headings,
-    };
-  });
-}
-
-/** Same matching rule as apps/docs reference/stats/[name] page load. */
-function statHasRelatedExamples(stat: string): boolean {
-  const compact = stat.replaceAll("_", "");
-  const dashed = stat.replaceAll("_", "-");
-  return EXAMPLES.some(
-    (entry) =>
-      entry.category === stat ||
-      entry.category === compact ||
-      entry.tags.includes(stat) ||
-      entry.tags.includes(dashed) ||
-      entry.tags.includes(`stat-${stat}`) ||
-      entry.tags.includes(`stat_${stat}`),
-  );
-}
-
-/** One indexable page per KNOWN_STATS entry. */
-function statDetailRoutes(): DocsRouteRecord[] {
-  return statReferenceList().map((entry) => {
-    const headings: RouteHeading[] = [
-      { id: "usage", title: "Usage", level: 2 },
-      { id: "generated-columns", title: "Generated columns (after_stat)", level: 2 },
-      { id: "default-for", title: "Default for geoms", level: 2 },
-      { id: "compatible-geoms", title: "Compatible geoms", level: 2 },
-    ];
-    if (statHasRelatedExamples(entry.name)) {
-      headings.push({ id: "examples", title: "Examples", level: 2 });
-    }
-    return {
-      path: `/reference/stats/${entry.slug}`,
-      title: `stat ${entry.name} — ggsvelte`,
-      description: `stat "${entry.name}": ${entry.summary}`,
-      canonicalPath: `/reference/stats/${entry.slug}`,
-      kind: "page" as const,
-      index: true,
-      sitemap: true,
-      shell: "docs" as const,
-      headings,
-    };
-  });
-}
-
-/** Same matching rule as apps/docs reference/positions/[name] page load. */
-function positionHasRelatedExamples(position: string): boolean {
-  return EXAMPLES.some(
-    (entry) =>
-      entry.tags.includes(position) ||
-      entry.tags.includes(`position-${position}`) ||
-      entry.tags.includes(`position_${position}`) ||
-      entry.id.includes(position),
-  );
-}
-
-/** One indexable page per KNOWN_POSITIONS entry. */
-function positionDetailRoutes(): DocsRouteRecord[] {
-  return positionReferenceList().map((entry) => {
-    const headings: RouteHeading[] = [
-      { id: "usage", title: "Usage", level: 2 },
-      { id: "params", title: "positionParams", level: 2 },
-      { id: "default-for", title: "Default for geoms", level: 2 },
-      { id: "compatible-geoms", title: "Compatible geoms", level: 2 },
-    ];
-    if (positionHasRelatedExamples(entry.name)) {
-      headings.push({ id: "examples", title: "Examples", level: 2 });
-    }
-    return {
-      path: `/reference/positions/${entry.slug}`,
-      title: `position ${entry.name} — ggsvelte`,
-      description: `position "${entry.name}": ${entry.summary}`,
-      canonicalPath: `/reference/positions/${entry.slug}`,
-      kind: "page" as const,
-      index: true,
-      sitemap: true,
-      shell: "docs" as const,
-      headings,
-    };
-  });
-}
-
-const ENDPOINT_ROUTES: readonly DocsRouteRecord[] = [
-  ["/schema/v0.json", "PortableSpec JSON Schema", "Machine-readable ggsvelte PortableSpec schema."],
-  ["/llms.txt", "ggsvelte documentation index", "A concise machine-readable documentation index."],
-  [
-    "/llms-full.txt",
-    "Complete ggsvelte documentation",
-    "Complete guide and example source for machine readers.",
-  ],
-  ["/sitemap.xml", "Sitemap", "Generated canonical route sitemap."],
-  ["/robots.txt", "Robots policy", "Crawler policy and absolute sitemap location."],
-].map(([path, title, description]) => ({
-  path: path!,
-  title: title!,
-  description: description!,
-  canonicalPath: path!,
-  kind: "endpoint",
-  index: false,
-  sitemap: false,
-  shell: "site",
-}));
-
-const PERFORMANCE_ROUTES: readonly DocsRouteRecord[] = [
-  "/__perf/r3-interaction",
-  "/__perf/interaction-100k",
-  "/__perf/legend-focus-100k",
-].map((path) => ({
-  path,
-  title: "Performance fixture — ggsvelte",
-  description: "Internal ggsvelte performance fixture; not part of the public documentation.",
-  canonicalPath: path,
-  kind: "performance",
-  index: false,
-  sitemap: false,
-  shell: "site",
-}));
 
 /** SEO meta needs a non-empty string; page lede may be empty (description deleted). */
 function exampleSeoDescription(entry: {
